final_project/requirements_venv.txt
absl-py==1.4.0
anyio==3.6.2
astunparse==1.6.3
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
fastai==2.7.11
fastapi==0.95.0
fastcore==1.5.28
fastdownload==0.0.7
fastprogress==1.0.3
flatbuffers==23.3.3
fonttools==4.39.2
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kaggle==1.5.13
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
murmurhash==1.0.9
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.26.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
unicorn==2.0.1.post1
urllib3==1.26.15
uvicorn==0.21.1
wasabi==1.1.1
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
